move most of the stuff from lib.rs into machine.rs, and initialization + main loop into eval.rs
